+++ src/sn/CmnInterface.ts
@@ -89,6 +89,10 @@ export type HArg = {
 	style?			: string;
 	style_hover?	: string;
 	style_clicked?	: string;
+	r_style?		: string;
+	r_style_hover?	: string;
+	r_style_clicked?: string;
+	r_align?	: string;
 
 	b_width?	: string;
 	b_height?	: string;
@@ -98,7 +102,6 @@ export type HArg = {
 	b_pic?		: string;
 	back_clear?	: string;
 
-	r_align?	: string;
 	max_col?	: string;
 	max_row?	: string;
 	bura_col?	: string;
@@ -114,7 +117,6 @@ export type HArg = {
 	breakout?	: Function;
 	arg?	: HArg;
 	fnc?	: (e: Event)=> void;
-	fold?	: boolean;
 
 	filter?	: string;